Now mod can run on servers
