+++ forge/src/main/java/net/yeoxuhang/geodeplus/forge/GeodePlusForge.java
@@ -4,6 +4,7 @@ import com.mojang.datafixers.util.Pair;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.yeoxuhang.geodeplus.GeodePlus;
+import net.yeoxuhang.geodeplus.client.GeodeClient;
 import net.yeoxuhang.geodeplus.client.render.WrappistPedestalBlockEntityRenderer;
 import net.yeoxuhang.geodeplus.common.registry.GeodeModBlockEntitiesRegistry;
 import net.yeoxuhang.geodeplus.forge.registry.GeodeModBiomeModifiersRegistry;
@@ -45,7 +46,7 @@ public class GeodePlusForge {
 
         @SubscribeEvent
         public static void clientSetup(FMLClientSetupEvent e) {
-            GeodePlus.initClient();
+            GeodeClient.initClient();
             BlockEntityRenderers.register(GeodeModBlockEntitiesRegistry.WRAPPIST_PEDESTAL_ENTITY.get(), WrappistPedestalBlockEntityRenderer::new);
         }
